pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Михаил Касаткин <m.m.kasatkin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
django = "^4.2.5"
python-environ = "^0.4.54"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.7"
tqdm = "^4.66.1"
django-filter = "^23.3"
drf-yasg = "^1.21.7"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^23.9.1"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
